backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
sqlalchemy==2.0.23
markdown2==2.4.10
weasyprint==60.1
openai==1.3.5
python-dotenv==1.0.0
validators==0.22.0
ratelimit==2.2.1
tiktoken

